refactor(chat): Modulize ChatPanel, fix loading issue when analyze the response
